==> _competition/2016/07_11_talkingdata-mobile-user-demographics/05_Combining_values.ipynb <==
Just took results from various scripts and aggregated them. What I took:

 - [Linear model on apps and labels](https://www.kaggle.com/dvasyukova/talkingdata-mobile-user-demographics/a-linear-model-on-apps-and-labels/notebook) **2.27**
 - [keras bag of apps](https://www.kaggle.com/alexxanderlarko/talkingdata-mobile-user-demographics/bag-of-apps-keras-11-08-16/code) **?**
 - [bag of apps](https://www.kaggle.com/xiaoml/talkingdata-mobile-user-demographics/bag-of-app-id-python-2-27392) **2.27**
 - [xgboost sparse](https://www.kaggle.com/anokas/talkingdata-mobile-user-demographics/sparse-xgboost-starter-2-26857) **2.26**
 - [keras labels and brands](https://www.kaggle.com/chechir/talkingdata-mobile-user-demographics/keras-on-labels-and-brands/code) **2.25**
 - [brands and models](https://www.kaggle.com/dvasyukova/talkingdata-mobile-user-demographics/brand-and-model-based-benchmarks/notebook) **2.39**

import pandas as pd
import numpy as np

x1 = pd.DataFrame.from_csv("data/res/1_apps.csv")
x2 = pd.DataFrame.from_csv("data/res/2_apps.csv")
x3 = pd.DataFrame.from_csv("data/res/3_apps.csv")
x4 = pd.DataFrame.from_csv("data/res/4_xgb.csv")
x5 = pd.DataFrame.from_csv("data/res/5_brands.csv") # looks very strange prob do not sum to 1
x6 = pd.DataFrame.from_csv("data/res/6_brands.csv")

x_ = pd.DataFrame.from_csv("train.csv", index_col='device_id')[['e_num']]
no_events = set(x_[x_['e_num'] == 0].index.values)

res = pd.concat([x1, x2, x3, x4, x6]).groupby(level=0).mean()
res.to_csv("output_1.csv")

res = pd.concat([x1, x2, x3, x4]).groupby(level=0).mean()
res.to_csv("output_2.csv")
